messaging_app/messaging_app/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
flake8==7.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mysqlclient==2.2.7
packaging==24.2
parameterized==0.9.0
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
